Move default Convertibles to a new file
This will make it easier to make a subspec that excludes these.
